pyproject.toml
[project]
name = "cash-llm-sample"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "takupista", email = "53781582+takupista@users.noreply.github.com" }
]
dependencies = [
    "pydantic>=2.5.2",
    "google-api-python-client>=2.108.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "openai>=1.3.7",
    "email-validator>=2.1.0.post1",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.345",
    "langchain-experimental>=0.0.43",
    "pandas>=2.1.3",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
